remove FileShareActivity
