+++ demo_app/src/main/java/com/bandyer/demo_sdk_design/MainActivity.kt
@@ -16,18 +16,26 @@
 
 package com.bandyer.demo_sdk_design
 
+import android.content.Context
 import android.content.Intent
 import android.content.res.Configuration
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import android.view.MotionEvent
 import android.widget.Toast
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
+import androidx.core.net.toUri
 import com.bandyer.demo_sdk_design.databinding.ActivityMainBinding
 import com.bandyer.sdk_design.bottom_sheet.items.ActionItem
 import com.bandyer.sdk_design.call.bottom_sheet.items.CallAction
 import com.bandyer.sdk_design.call.dialogs.BandyerSnapshotDialog
+import com.bandyer.sdk_design.filesharing.BandyerFileShareDialog
+import com.bandyer.sdk_design.filesharing.FileShareViewModel
+import com.bandyer.sdk_design.filesharing.model.FileInfo
+import com.bandyer.sdk_design.filesharing.model.FileShareItemData
 import com.bandyer.sdk_design.smartglass.call.menu.SmartGlassActionItemMenu
 import com.bandyer.sdk_design.smartglass.call.menu.SmartGlassMenuLayout
 import com.bandyer.sdk_design.smartglass.call.menu.items.getSmartglassActions
@@ -35,6 +43,7 @@ import com.bandyer.sdk_design.smartglass.call.menu.utils.MotionEventInterceptor
 import com.bandyer.sdk_design.whiteboard.dialog.BandyerWhiteboardTextEditorDialog
 import com.bandyer.sdk_design.whiteboard.dialog.BandyerWhiteboardTextEditorDialog.BandyerWhiteboardTextEditorWidgetListener
 import com.google.android.material.appbar.MaterialToolbar
+import java.util.concurrent.ConcurrentHashMap
 
 class MainActivity : AppCompatActivity() {
 
@@ -46,6 +55,8 @@ class MainActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityMainBinding
 
+    private val viewModel: LocalFileShareViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
@@ -53,6 +64,15 @@ class MainActivity : AppCompatActivity() {
 
         setSupportActionBar(findViewById<MaterialToolbar>(R.id.toolbar))
         initializeListeners()
+        initFileShareItems()
+    }
+
+    private fun initFileShareItems() {
+        viewModel.itemsData["id_1"] = FileShareItemData(FileInfo(id = "1", uri = "".toUri(), name = "razer.jpg", mimeType = "image/jpeg", sender = "Gianluigi", size = 100L), FileShareItemData.State.Pending, FileShareItemData.Type.DownloadAvailable)
+        viewModel.itemsData["id_2"] = FileShareItemData(FileInfo(id = "2", uri = "".toUri(), name = "identity_card.pdf", mimeType = "", sender = "Mario", size = 100L), FileShareItemData.State.Success("".toUri()), FileShareItemData.Type.Download)
+        viewModel.itemsData["id_3"] = FileShareItemData(FileInfo(id = "3", uri = "".toUri(), name = "car.zip", mimeType = "application/zip", sender = "Luigi", size = 1000L), FileShareItemData.State.OnProgress(600L), FileShareItemData.Type.Download)
+        viewModel.itemsData["id_4"] = FileShareItemData(FileInfo(id = "4", uri = "".toUri(), name = "phone.doc", mimeType = "", sender = "Gianni", size = 23000000L), FileShareItemData.State.Error(Throwable()), FileShareItemData.Type.Upload)
+        viewModel.itemsData["id_5"] = FileShareItemData(FileInfo(id = "5", uri = "".toUri(), name = "address.jpg", mimeType = "image/jpeg", sender = "Marco", size = 1000L), FileShareItemData.State.Pending, FileShareItemData.Type.Upload)
     }
 
     private fun initializeListeners() {
@@ -96,7 +116,7 @@ class MainActivity : AppCompatActivity() {
 
         binding.btnBluetoothAudioroute.setOnClickListener { startActivity(Intent(this, BluetoothAudioRouteActivity::class.java)) }
 
-        binding.btnFileShare.setOnClickListener { startActivity(Intent(this, FileShareActivity::class.java)) }
+        binding.btnFileShare.setOnClickListener { BandyerFileShareDialog().show(this@MainActivity, viewModel, viewModel.itemsData) }
     }
 
     override fun dispatchTouchEvent(ev: MotionEvent?): Boolean {
@@ -125,4 +145,16 @@ class MainActivity : AppCompatActivity() {
                     }
                 }
             }
+}
+
+class LocalFileShareViewModel: FileShareViewModel() {
+    val itemsData: ConcurrentHashMap<String, FileShareItemData> = ConcurrentHashMap()
+
+    override fun upload(context: Context, id: String, uri: Uri, sender: String) = Unit
+
+    override fun download(context: Context, id: String, uri: Uri, sender: String) = Unit
+
+    override fun cancelUpload(uploadId: String) = Unit
+
+    override fun cancelDownload(downloadId: String) = Unit
 }
